titan-poss-poss-web-f78a2c9ef984/titan-poss-poss-web-f78a2c9ef984/libs/eposs/max-flat-tep-config/data-access-max-flat-tep-config/src/lib/+state/max-flat-tep-config.selectors.ts
import { createSelector } from '@ngrx/store';
import { selectMaxFlatTepConfigState } from './max-flat-tep-config.reducer';

const selectError = createSelector(
  selectMaxFlatTepConfigState,
  state => state.errors
);

const selectIsLoading = createSelector(
  selectMaxFlatTepConfigState,
  state => state.isLoading
);

const selectMaxFlatTepConfigDetails = createSelector(
  selectMaxFlatTepConfigState,
  state => state.maxFlatTepConfigDetails
);

const selectUpdateMaxFlatTepConfigResponse = createSelector(
  selectMaxFlatTepConfigState,
  state => state.updateMaxFlatTepConfigResponse
);

export const MaxFlatTepConfigSelectors = {
  selectError,
  selectIsLoading,
  selectMaxFlatTepConfigDetails,
  selectUpdateMaxFlatTepConfigResponse
};
